+++ applitools.config.js
@@ -11,10 +11,9 @@ function getHeadHash() {
     .trim();
 }
 
-process.env.APPLITOOLS_DONT_CLOSE_BATCHES = true;
-
 module.exports = {
   ...privateConfig,
   concurrency: 200,
   batchId: getHeadHash(),
+  dontCloseBatches: true,
 };