+++ packages/editor-common/web/src/Base/createBasePlugin.jsx
@@ -36,7 +36,6 @@ const createBasePlugin = (config = {}, underlyingPlugin) => {
     customStyleFn,
     getEditorBounds,
     onOverlayClick,
-    onAtomicBlockFocus,
     disableRightClick,
     commonPubsub,
     defaultPluginData,
@@ -91,7 +90,6 @@ const createBasePlugin = (config = {}, underlyingPlugin) => {
       pluginDecorationProps: config.pluginDecorationProps,
       componentWillReceiveDecorationProps: config.componentWillReceiveDecorationProps,
       onOverlayClick,
-      onAtomicBlockFocus,
       pubsub,
       commonPubsub,
       settings,
@@ -111,7 +109,7 @@ const createBasePlugin = (config = {}, underlyingPlugin) => {
 
   const TextButtonMapper = config.toolbar && config.toolbar.TextButtonMapper;
 
-  const blockRendererFn = (contentBlock, { getEditorState, setEditorState, getReadOnly }) => {
+  const blockRendererFn = (contentBlock, { getEditorState, setEditorState }) => {
     if (contentBlock.getType() === 'atomic') {
       // TODO subject to change for draft-js next release
       const contentState = getEditorState().getCurrentContent();
@@ -131,7 +129,6 @@ const createBasePlugin = (config = {}, underlyingPlugin) => {
                 setEditorState,
               }),
               deleteBlock: deleteEntity(contentBlock, { getEditorState, setEditorState }),
-              readOnly: getReadOnly(),
             },
           };
         }