+++ packages/editor-common/web/src/index.js
@@ -61,6 +61,7 @@ export {
   insertLinkAtCurrentSelection,
   insertLinkInPosition,
   hasLinksInBlock,
+  fixPastedLinks,
   hasLinksInSelection,
   getLinkDataInSelection,
   removeLinksInSelection,
@@ -73,6 +74,9 @@ export {
   replaceWithEmptyBlock,
   deleteBlock,
   createEntity,
+  createBlock,
+  getBlockInfo,
+  getFocusedBlockKey,
 } from './Utils/draftUtils';
 export { isiOS } from './Utils/isiOS';
 export { getSelectionStyles } from './Utils/inlineStyleUtils';