+++ packages/plugin-emoji/web/src/icons/index.js
@@ -1,2 +1,1 @@
-export { default as EmojiPluginIcon } from './EmojiPluginIcon';
 export { default as JoyPixelsIcon } from './JoyPixelsIcon';