+++ CHANGELOG.md
@@ -13,9 +13,42 @@
     Click to see more.
   </summary>
 
+  ### :bug: Bug Fix
+  - `editor-common`
+    - [#550](https://github.com/wix-incubator/rich-content/pull/550) remove dashed border when focused
+    - [#565](https://github.com/wix-incubator/rich-content/pull/565) undo remembers all plugins changes
+  - `link`
+    - [#546](https://github.com/wix-incubator/rich-content/pull/546) saves the last data and the initial state of the checkboxes("Open in a new tab", "Add a nofollow tag") is according to the defaults (anchorTarget, relValue)
+  ### :house: Internal
+  - `general`
+    - [#561](https://github.com/wix-incubator/rich-content/pull/561) refactor - cleanup readOnly prop from code
+    - [#562](https://github.com/wix-incubator/rich-content/pull/562) use `@lerna/project` to get all lerna packages
+  - `e2e`
+    - [#564](https://github.com/wix-incubator/rich-content/pull/564) snapshots renaming for gallery and image plugins tests
+
+</details>
+<hr/>
+
+## 6.6.5 (Dec 30, 2019)
+### :bug: Bug Fix
+- `editor-common`
+  - [#560](https://github.com/wix-incubator/rich-content/pull/560) getBlockInfo returns entity type rather block type
+
+## 6.6.4 (Dec 30, 2019)
+### :bug: Bug Fix
+- `editor-common`
+  - [#558](https://github.com/wix-incubator/rich-content/pull/558) getBlockInfo util arguments fixed
+
+## 6.6.3 (Dec 29, 2019)
   ### :rocket: New Feature
-  - `fullscreen`
-    - [#519](https://github.com/wix-incubator/rich-content/pull/519) add props: topMargin, backgroundColor, foregroundColor. Remove counter when only 1 image. Larger z-index
+   - `text-color`
+     - [#520](https://github.com/wix-incubator/rich-content/pull/520) color Picker- adding an option to change back to default color
+  ### :bug: Bug Fix
+   - `editor`
+      - [#556](https://github.com/wix-incubator/rich-content/pull/556) fix onAtomicBlockFocus to work when alternating focus between different kinds of plugins
+
+## 6.6.2 (Dec 11, 2019)
+ ### :rocket: New Feature
   - `editor-common`
     - [#526](https://github.com/wix-incubator/rich-content/pull/526) `onAtomicBlockFocus` is invoked with `undefined` when an atomic block loses focus
     - [#530](https://github.com/wix-incubator/rich-content/pull/530) `uiSettings.linkPanel.placeholder` determines the link panel's placeholder
@@ -24,11 +57,18 @@
   - `file-upload`
     - [#533](https://github.com/wix-incubator/rich-content/pull/533) add multiple files at once
   ### :bug: Bug Fix
-  - `divider`
-    - [#511](https://github.com/wix-incubator/rich-content/pull/511) divider's width set to 100% in editor&viewer when no initial state of width is given
+  - `editor-common`
+    - [#550](https://github.com/wix-incubator/rich-content/pull/550) remove dashed border when focused
   - `editor`
+    - [#504](https://github.com/wix-incubator/rich-content/pull/504) pasted links get the configured anchorTarget and relValue
     - [#521](https://github.com/wix-incubator/rich-content/pull/521) replace p element with div element in lists
-    - [#565](https://github.com/wix-incubator/rich-content/pull/565) undo remembers all plugins changes
+    - [#552](https://github.com/wix-incubator/rich-content/pull/552) fix Plugin toolbar shouldCreate config crush
+  - `editor`
+    - [#554](https://github.com/wix-incubator/rich-content/pull/554) change align shortcuts from cmd to cmd + shift
+  - `undo-redo`
+    - [#543](https://github.com/wix-incubator/rich-content/pull/543) fix undo-redo icons on mobile
+  - `gallery`
+    - [#545](https://github.com/wix-incubator/rich-content/pull/545) fix image ratio change in gallery slider layout
   ### :house: Internal
   - `gallery-settings`
     - [#535](https://github.com/wix-incubator/rich-content/pull/535) gallery setting's header, tabs and footer position
@@ -38,7 +78,7 @@
     - [#522](https://github.com/wix-incubator/rich-content/pull/522) support legacy margin scss attributes
   - `e2e`
     - [#538](https://github.com/wix-incubator/rich-content/pull/538) moved tests out of Docker, e2e and unit tests run in github actions
-  ### :house: Internal
+    - [#542](https://github.com/wix-incubator/rich-content/pull/542) e2e tests run visual tests and content snapshots separately
   - `codeBlock`
     - [#527](https://github.com/wix-incubator/rich-content/pull/527) remove onTab (moved to keyBindingFn/handleKeyCommand)
   - `pubsub`
@@ -47,8 +87,6 @@
     - [#534](https://github.com/wix-incubator/rich-content/pull/534) exposes data normalization and data validation utils as separate entries
   - `button`
     - [#517](https://github.com/wix-incubator/rich-content/pull/517) plugin button refactor
-</details>
-<hr/>
 
 ## 6.6.1 (Dec 11, 2019)
 ### :rocket: New Feature